refactor: 폴더 및 파일 이름 형식 통일
- domain을 제외한 폴더 이름은 복수형으로 사용
- 파일명은 camelCase가 아니라 kebab-case로 사용

ex. dto/createUser.dto.ts → dtos/create-user.dto.ts
