+++ backend/src/auth/guards/oauth-data.guard.ts
