+++ backend/src/dogs/dtos/dog-statistic.dto.ts
