+++ backend/src/dogs/guards/auth-dog.guard.ts
