+++ backend/src/journals/dtos/journal-list.dto.ts
