+++ backend/src/journals/dtos/journals-detail.dto.ts
