+++ backend/src/journals/guards/auth-journal.guard.ts
