+++ backend/src/journals/journals.controller.ts
@@ -1,11 +1,11 @@
 import { Body, Controller, Delete, Get, Param, ParseIntPipe, Patch, Post, Query, UseGuards } from '@nestjs/common';
 import { AccessTokenPayload } from 'src/auth/token/token.service';
-import { DateValidationPipe } from 'src/statistics/pipes/dateValidation.pipe';
+import { DateValidationPipe } from 'src/statistics/pipes/date-validation.pipe';
 import { User } from 'src/users/decorators/user.decorator';
-import { JournalInfoForList } from './dto/journal-list.dto';
-import { UpdateJournalDto } from './dto/journal-update.dto';
-import { CreateJournalDto } from './dto/journals-create.dto';
-import { AuthJournalGuard } from './guards/authJournal.guard';
+import { JournalInfoForList } from './dtos/journal-list.dto';
+import { UpdateJournalDto } from './dtos/journal-update.dto';
+import { CreateJournalDto } from './dtos/journals-create.dto';
+import { AuthJournalGuard } from './guards/auth-journal.guard';
 import { JournalsService } from './journals.service';
 
 @Controller('/journals')
