+++ backend/src/s3/types/s3.type.ts
